requirements.txt
playwright==1.44.0
pytest==7.4.0
pytest-playwright==0.6.0
allure-pytest==2.10.0
python-dotenv==1.0.0

allure-python-commons